pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jewei-demo-mcp-server"
version = "0.1.0"
authors = [
    {name = "Jewei", email = "jiweifong@qq.com"}
]
description = "DEMO MCP服务器，提供DEMO功能"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastmcp>=0.1.0",
    "python-dotenv>=0.19.0",
]

[project.urls]
"Homepage" = "https://github.com/jeweis/jewei-demo-mcp-server"
"Bug Tracker" = "https://github.com/jeweis/jewei-demo-mcp-server/issues"

[project.scripts]
jewei-demo-mcp-server = "jewei_demo_mcp_server.server:main"

[tool.setuptools]
packages = ["jewei_demo_mcp_server"]